requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
psycopg[binary]==3.1.18
sqlalchemy==2.0.29
python-dotenv==1.0.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
beautifulsoup4==4.12.3
email-validator==2.1.0
